requirements.txt
streamlit
openai
python-dotenv
pandas
matplotlib
pypdf
openpyxl
tabulate